Add compressedRefs support
Creates two memory regions, a <4gb region and a full range region. In
the future this will be updated to support a dynamic range of memory
regions. The mapping code has been modified to map regions individually.
mmap requires the file offsets for mapping to be page aligned, as a
result padding between regions is added to meet these constraints.

Also, the unused fields in the snapshot header have been removed.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>