Update MethodHandleResolver
1. In OJDK, MHN.linkMethodHandleConstant resolves a MethodHandle
constant pool entry. It is being incorporated in OpenJ9 to be used with
OJDK MHs. It has certain fast-paths which improve the performance.

2. Add inaccessible type checks in OJDK MHs' version of
sendResolveMethodHandle and resolveInvokeDynamic. This is equivalent to
accessCheckArgRetTypes. inaccessibleTypeCheck is a static version of
accessCheckArgRetTypes.

3. Convert Lookup.isClassAccessible from a private to package protected
method. inaccessibleTypeCheck uses this method. This change is performed
to use inaccessibleTypeCheck in OpenJ9

4. Remove the try block from OJDK MH's version of resolveInvokeDynamic.
It is not needed for Java 16. It incorrectly converts certain throwables
to different ones. Also, MethodHandleNatives.linkCallSite throws the
correct throwable. So, the throwable does not need to be caught and
converted into a different one. For JDK8, this method may need to be
re-visited due to spec differences. Currently, the spec differences
cannot be identified since OpenJ9 JDK8 cannot be built with OJDK MHs
enabled.

Fixes: https://github.com/eclipse/openj9/issues/11934

Related: https://github.com/eclipse/openj9/issues/10913

Co-authored-by: Jack Lu <Jack.S.Lu@ibm.com>
Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>