Update error handling in JVMTI NotifyFramePop
Return JVMTI_ERROR_THREAD_NOT_SUSPENDED if the thread was not
suspended and was not the current thread.

For virtual threads, VirtualThread.state is read and it is checked
if JVMTI_VTHREAD_STATE_SUSPENDED is set in VirtualThread.state.

This check is incorrect because JVMTI_VTHREAD_STATE_SUSPENDED
represents an internal state at the Java level. It does not
imply if the virtual thread is suspended through JVMTI.

JVMTI NotifyFramePop is updated to check if the virtual thread
is suspended through JVMTI.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>